# [3][Performance] Query airtable by city