--- a/backend/app.py
+++ b/backend/app.py
@@ -170,8 +170,8 @@ async def government_paid_hospital_data() -> List[Hospital]:
 # pyre-fixme[56]: Decorator async types are not type-checked.
 @app.route("/government_paid_hospitals")
 async def government_paid_hospitals() -> wrappers.Response:
-    print(request.args, file=sys.stdout)
-    data = await get_hospitals_from_airtable()
+    city: str = request.args.get("city")
+    data = await get_hospitals_from_airtable("", city)
     response = app.response_class(
         response=json.dumps(data),
         status=200,