--- a/backend/hospital_types.py
+++ b/backend/hospital_types.py
@@ -17,7 +17,7 @@ class Hospital(TypedDict):
     governmentPaidAvailability: AppointmentAvailability
     hospitalId: HospitalID
     location: str
-    county: str
+    district: str
     name: str
     phone: str
     selfPaidAvailability: AppointmentAvailability