--- a/frontend/Components/VaccineInfo/Card.jsx
+++ b/frontend/Components/VaccineInfo/Card.jsx
@@ -34,7 +34,7 @@ export default function Card(props: {
   buttonText: string,
   department: string,
   location: Location,
-  county: string,
+  district: string,
   name: string,
   phone: string,
   website: string,
@@ -45,7 +45,7 @@ export default function Card(props: {
     buttonText,
     department,
     location,
-    county,
+    district,
     name,
     phone,
     website,
@@ -63,19 +63,26 @@ export default function Card(props: {
           <span className="badge bg-dark text-light me-1">
             {getLocationName(location, cityT)}
           </span>
-          <span className="badge bg-dark text-light me-1">
-            {county}
-          </span>
+          <span className="badge bg-dark text-light me-1">{district}</span>
         </p>
         <h4 className="card-title">{name}</h4>
         <h6 className="card-subtitle mb-2 text-muted">{address}</h6>
         <p className="card-text">{department}</p>
         <p className="card-text">{phone}</p>
         <div className="d-grid mt-auto">
-          {
-            website != null ? <a href={website} className="btn btn-primary mb-1" target="_blank" rel="noreferrer">{buttonText}</a> : null
-          }
-          <a href={`tel:${phone}`} className="btn btn-primary mb-1 d-md-none">電話預約</a>
+          {website != null ? (
+            <a
+              href={website}
+              className="btn btn-primary mb-1"
+              target="_blank"
+              rel="noreferrer"
+            >
+              {buttonText}
+            </a>
+          ) : null}
+          <a href={`tel:${phone}`} className="btn btn-primary mb-1 d-md-none">
+            電話預約
+          </a>
         </div>
       </div>
     </div>