--- a/frontend/Components/VaccineInfo/Cards.jsx
+++ b/frontend/Components/VaccineInfo/Cards.jsx
@@ -6,9 +6,11 @@ import { getAvailability } from '../../Types/Hospital';
 import type { Hospital } from '../../Types/Hospital';
 import type { VaccineType } from '../../Types/VaccineType';
 
-export default function Cards(
-  props: { hospitals: Hospital[], vaccineType: VaccineType, buttonText: string},
-): React.Node {
+export default function Cards(props: {
+  hospitals: Hospital[],
+  vaccineType: VaccineType,
+  buttonText: string,
+}): React.Node {
   const { hospitals, buttonText, vaccineType } = props;
 
   return hospitals.map((hospital) => (
@@ -19,7 +21,7 @@ export default function Cards(
         buttonText={buttonText}
         department={hospital.department}
         location={hospital.location}
-        county={hospital.county}
+        district={hospital.district}
         name={hospital.name}
         phone={hospital.phone}
         website={hospital.website}