--- a/frontend/Components/VaccineInfo/DataGrid.jsx
+++ b/frontend/Components/VaccineInfo/DataGrid.jsx
@@ -5,51 +5,45 @@ import * as React from 'react';
 import Cards from './Cards';
 
 import type { Hospital } from '../../Types/Hospital';
-import type { Location } from '../../Types/Location';
 import type { VaccineType } from '../../Types/VaccineType';
 
 export default function DataGrid(props: {
   hospitals: Hospital[],
   buttonText: string,
   vaccineType: VaccineType,
-  selectedLocation: string,
-  selectedCounty: string
+  selectedCounty: string,
 }): React.Node {
   const {
-    hospitals, buttonText, vaccineType, selectedLocation, selectedCounty,
+    hospitals, buttonText, vaccineType, selectedCounty,
   } = props;
 
-  const hospitalsByCity = hospitals.reduce((byCity: { [Location]: Hospital[] }, hospital) => {
-    if (hospital.location in byCity) {
-      byCity[hospital.location].push(hospital);
-      return byCity;
-    }
-
-    const newLocation = {};
-    newLocation[hospital.location] = [hospital];
-    return { ...byCity, ...newLocation };
-  }, {});
-
   if (hospitals.length === 0) {
-    return (
-      <div> </div>
-    );
+    return <div> </div>;
   }
 
-  const makeCardGrid: (Hospital[]) =>
-  React.Node = (localHospitals) => (
+  const makeCardGrid: (Hospital[]) => React.Node = (localHospitals) => (
     <div className="row row-cols-1 row-cols-md-3 g-3">
-      <Cards hospitals={localHospitals} buttonText={buttonText} vaccineType={vaccineType} />
+      <Cards
+        hospitals={localHospitals}
+        buttonText={buttonText}
+        vaccineType={vaccineType}
+      />
     </div>
   );
 
-  return (hospitals.length <= 20 ? makeCardGrid(hospitals)
-    : (
-      <>
-        {/* $FlowFixMe: Casting from enum to string. */}
-        {/* $FlowFixMe: Casting from a string to an Enum. */}
-        {makeCardGrid(hospitalsByCity[selectedLocation].filter((hospital) => hospital !== undefined).filter((hospital) => (selectedCounty === 'null') || (hospital.county === selectedCounty)))}
-      </>
-    )
+  return hospitals.length <= 20 ? (
+    makeCardGrid(hospitals)
+  ) : (
+    <>
+      {/* $FlowFixMe: Casting from enum to string. */}
+      {/* $FlowFixMe: Casting from a string to an Enum. */}
+      {makeCardGrid(
+        hospitals
+          .filter((hospital) => hospital !== undefined)
+          .filter(
+            (hospital) => selectedCounty === 'null' || hospital.county === selectedCounty,
+          ),
+      )}
+    </>
   );
 }