--- a/frontend/Types/Hospital.js
+++ b/frontend/Types/Hospital.js
@@ -3,20 +3,24 @@ import type { Availability } from './Availability';
 import type { Location } from './Location';
 import type { VaccineType } from './VaccineType';
 
-export type Hospital = {
-    address: string,
-    department: string,
-    governmentPaidAvailability: Availability,
-    hospitalId: number,
-    location: Location,
-    county: string,
-    name: string,
-    phone: string,
-    selfPaidAvailability: Availability,
-    website: string,
-  };
+export type Hospital = {|
+  address: string,
+  department: string,
+  governmentPaidAvailability: Availability,
+  hospitalId: number,
+  location: Location,
+  district: string,
+  name: string,
+  phone: string,
+  selfPaidAvailability: Availability,
+  website: string,
+|};
 
-export function getAvailability(hospital: Hospital, vaccineType: VaccineType): Availability {
+export function getAvailability(
+  hospital: Hospital,
+  vaccineType: VaccineType,
+): Availability {
   return vaccineType === 'SelfPaid'
-    ? hospital.selfPaidAvailability : hospital.governmentPaidAvailability;
+    ? hospital.selfPaidAvailability
+    : hospital.governmentPaidAvailability;
 }